abstracting form from register component
